diopsis-submission/requirements.txt
pandas~=2.1.4
scikit-learn~=1.3.2
numpy~=1.26.3
matplotlib~=3.8.2
torch
torchvision
pillow
